+++ .github/scripts/build-posix.sh
@@ -40,6 +40,8 @@ export PKG_CONFIG_PATH=/usr/share/pkgconfig/pkgconfig
 git clone https://github.com/savonet/Camomile.git
 cd Camomile && opam install -y .
 
+opam install -y tls ca-certs
+
 echo "::endgroup::"
 
 echo "::group::Checking out CI commit"