+++ src/core/tools/http.mli
@@ -11,12 +11,15 @@ type socket =
   ; read : Bytes.t -> int -> int -> int
   ; close : unit >
 
+and server =
+  < transport : transport ; accept : Unix.file_descr -> socket * Unix.sockaddr >
+
 and transport =
   < name : string
   ; protocol : string
   ; default_port : int
   ; connect : ?bind_address:string -> ?timeout:float -> string -> int -> socket
-  ; accept : Unix.file_descr -> socket * Unix.sockaddr >
+  ; server : server >
 
 type uri = {
   protocol : string;